refactor: examples renames to snippets, & examples.ts to http.ts
